requirements.txt
langchain
langchain-core
langchain-community
chromadb
google-generativeai
pymupdf
python-dotenv
tiktoken
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai
# Environment Variable Management
python-dotenv
# Machine Learning Utilities
numpy
scikit-learn

langchain-chroma
pdfplumber